* Removed mina- prefix from module directory names like other existing submodules * Cleaned up poms
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@575140 13f79535-47bb-0310-9956-ffa450edef68

+++ protocol-client-http/src/test/java/org/apache/mina/http/client/FakeIoSession.java
@@ -258,13 +258,13 @@ public class FakeIoSession implements IoSession
     }
 
 
-    public int getScheduledWriteRequests()
+    public long getScheduledWriteRequests()
     {
         return 0;
     }
 
 
-    public int getScheduledWriteBytes()
+    public long getScheduledWriteBytes()
     {
         return 0;
     }
